Fixed Unicode serialization of CString object.
[git-p4: depot-paths = "//depot/openwatcom/": change = 36433]

+++ bld/afxapi/lib/archive.cpp
@@ -308,7 +308,7 @@ CArchive &CArchive::operator<<( const CString &str )
     int nLength = str.GetLength();
 #ifdef _UNICODE
     (*this) << (BYTE)0xFF;
-    (*this) << (BYTE)0xFFFE;
+    (*this) << (WORD)0xFFFE;
 #endif
     if( nLength < 0xFF ) {
         (*this) << (BYTE)nLength;
